requirements.txt
requests>=2.28.0
python-dotenv>=0.21.0
openai>=1.0.0
schedule>=1.1.0 
pytz>=2023.3 